som_devrobs.c Fixed nasty bug in map_from_dataset().  Changed == to =.  Now, we can free last_coords in the while loop as well as before we return from the function.
Free last_coords before returning from train_fromdataset().

Initialize data entry pointers (last and entry) before entering the while loop.
Reason for this:
"
if a user has been so ignorant as to issue two
training requests, his program gets killed with a segfault in the
get_label routine.  The problem is once again garbage being returned
to the ptyhon level (the last entry was uninitialized).  Attached is
a patch to a least give a semi-sane response to the python level."
WKV 2003-07-31