Works now, as far as I know.  Weights no longer grow to infinity for any reasonable values of the various settings. Additional testing still would be good.  Some code cleanup, comment cleanup, and docstring creation would probably be a good idea at this point as well.
The problem was using self.ACTPRIME as the derivative of the activation function when it is actually the derivative of the activation function
in terms of the activation value.  Composing it with the activation function is what I thought it was.  I created a function
called actDeriv that is the true derivative of the activation function.  This is inconsistent with the way the rest of conx behaves,
but I couldn't get it to work easily by taking self.ACTPRIME( self.activationFunction(...) ).  I think it is important
to have the true derivative of the activation function defined as a matter of principle and convenience as well so perhaps
actDeriv should be moved to the Network class instead of one of its subclasses.